Separated view route controllers into separate file in relevant directory
