+++ socketsDemo/server.js
@@ -6,20 +6,6 @@ const path = require('path');
 const io = require('socket.io').listen(server);
 const exphbs = require('express-handlebars');
 
-// Init Express Middleware
-app.use(express.urlencoded({ extended: false }));
-app.use(express.json());
-app.use(express.static("public"));
-
-// Init express handlebars engine
-app.engine(
-  "handlebars",
-  exphbs({
-    defaultLayout: "main"
-  })
-);
-app.set("view engine", "handlebars");
-
 // Init chatroom object
 const rooms = [
     {
@@ -54,46 +40,22 @@ const roomNames = rooms.map(function(room) {
 
 });
 
-// Home get controller
-app.get('/:home?', function(req, res, next) {
+// Init Express Middleware
+app.use(express.urlencoded({ extended: false }));
+app.use(express.json());
+app.use(express.static("public"));
 
-    // Serve home view if route goes to root, "home", or "index"
-    if (req.params.home === 'home' || req.params.home === 'index' || !req.params.home) {
+// Init routers, passing stored chatroom data
+app.use(require('./routes/viewRoutes.js')(rooms, roomNames));
 
-        // Send list of rooms to home view
-        res.render("home", {roomNames});
-
-    }
-
-    // Otherwise fall through to chat room route controller
-    else {
-        next();
-    };
-
-})
-
-// Chatroom get controller
-app.get('/:chat', function(req, res, next) {
-
-    // Find index of relevant chatroom, if exists, otherwise will be -1
-    const requestedRoomIndex = roomNames.indexOf(roomNames.find((name) => name.roomName === req.params.chat));
-
-    // If the chatroom exists then the index will not be -1
-    if (requestedRoomIndex !== -1) {
-        
-        // Serve given chatroom view
-        res.render('chat', rooms[requestedRoomIndex]);
-    
-    }
-
-    // If chatroom doesn't exist, redirect to home
-    else {
-
-        res.render('home', {roomNames});
-
-    };
-
-})
+// Init express handlebars engine
+app.engine(
+  "handlebars",
+  exphbs({
+    defaultLayout: "main"
+  })
+);
+app.set("view engine", "handlebars");
 
 // Establish socket.io server connection
 io.on('connection', function(socket) {
